requirements.txt
sqlite3
pycryptodome
pywin32
requests
browser_cookie3